refactor: renomeia arquivos das barras laterais
Renomeia os arquivos dos componentes de layout para padronizar a nomenclatura em português.
Os componentes da barra lateral esquerda e direita passam a usar os nomes BarraFerramentasEsquerda.py e BarraFerramentasDireita.py, deixando a estrutura do módulo layout mais alinhada com os nomes das classes exportadas no __init__.py.
Também atualiza os imports necessários para usar os novos nomes dos arquivos e manter a aplicação funcionando com a nova organização.
